separate do_get and do_put
